+++ robotbase/src/main/java/frc/robot/vision/LimelightCamera.java
@@ -65,7 +65,7 @@ public class LimelightCamera implements CameraInterface {
           m_camera
             .getSettings()
             .withThrottle(LIMELIGHT.ENABLED_THERMAL_THROTTLE)
-            .withImuMode(ImuMode.SyncInternalImu)
+            .withImuMode(ImuMode.InternalImuExternalAssist)
             .save();
         })
       );
@@ -74,12 +74,7 @@ public class LimelightCamera implements CameraInterface {
     SmartDashboard.putBoolean("seeded", false);
   }
 
-  public Command handleAutoStartCommand(Orientation3d robotOrientation) {
-    return Commands.runOnce(() -> seedGyroMethod(robotOrientation));
-  }
-
   public void seedGyroMethod(Orientation3d robotOrientation) {
-    System.out.println("seeding camera");
     SmartDashboard.putBoolean("seeded", true);
     m_camera
       .getSettings()
@@ -89,7 +84,10 @@ public class LimelightCamera implements CameraInterface {
   }
 
   public void useInternalImu() {
-    m_camera.getSettings().withImuMode(ImuMode.SyncInternalImu).save();
+    m_camera
+      .getSettings()
+      .withImuMode(ImuMode.InternalImuExternalAssist)
+      .save();
   }
 
   @Override
